Introduce ffi_type, csymbol and ktap_cdata
ffi type is basic type used in ffi module, which has fixed size and alignment.

csymbol means symbol for c type, which is designed to represent all
possible c types in its format. It doesn't contain any real value or data.

ktap_cdata is the uniform cdata type in ktap. It is an instance of csymbol,
which binds a csymbol with the real data. It mainly exists in the vm stack.

Signed-off-by: Yicheng Qin <qycqycqycqycqyc@gmail.com>
Signed-off-by: Qingping Hou <qingping.hou@gmail.com>